Extract crdt.IndexTree from crdt.Tree
